requirements.txt
asgiref==3.2.7
Django==3.0.6
django-appconf==1.0.4
django-imagekit==4.0.2
pilkit==2.0
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
rarfile==3.1
six==1.15.0
sqlparse==0.3.1
